decrypt.rb finalized; files renamed
