+++ decrypted.txt
@@ -1,0 +1,1 @@
+what a fun day :)
