+++ encrypted.txt
@@ -1,1 +1,1 @@
-mqhzqjglkwgjrgg:)f
+odcgsxbtmjbrtub:)n
