+++ lib/decrypt.rb
@@ -2,12 +2,11 @@ require 'pry'
 require './lib/enigma_decrypt'
 require 'date'
 # ARGV == ["encrypted.txt", "decrypted.txt", key, date]
+# key = "02715", date = "040895"
 encrypted_message = File.open(ARGV[0], "r")
-original_message = File.open(ARGV[1], "w")
+decrypted_message = File.open(ARGV[1], "w")
 encrypted_text = encrypted_message.read
-encrypted_text = Decrypt.new.decrypt(encrypted_text, ARGV[2], ARGV[3])
-binding.pry 
-encrypted_message.write(encrypted_text[:encryption])
-encrypted_message.close
-
-puts "Created '#{ARGV[1]}' with the key #{encrypted_text[:key]} and date #{encrypted_text[:date]}"
+decrypted_text = Decrypt.new.decrypt(encrypted_text, ARGV[2], ARGV[3])
+decrypted_message.write(decrypted_text[:decryption])
+decrypted_message.close
+puts "Created '#{ARGV[1]}' with the key #{decrypted_text[:key]} and date #{decrypted_text[:date]}"
