+++ lib/enigma_encrypt.rb
